# scripts/openapi-snapshot.json
{
 "components": {
  "schemas": {
   "AcceptIn": {
    "properties": {
     "display_name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Display Name",
      "type": "string"
     },
     "password": {
      "maxLength": 1024,
      "minLength": 1,
      "title": "Password",
      "type": "string"
     },
     "token": {
      "maxLength": 200,
      "minLength": 20,
      "title": "Token",
      "type": "string"
     }
    },
    "required": [
     "token",
     "display_name",
     "password"
    ],
    "title": "AcceptIn",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "InvitationCreated": {
    "properties": {
     "accept_token": {
      "title": "Accept Token",
      "type": "string"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "email": {
      "title": "Email",
      "type": "string"
     },
     "expires_at": {
      "format": "date-time",
      "title": "Expires At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "role": {
      "title": "Role",
      "type": "string"
     }
    },
    "required": [
     "id",
     "email",
     "role",
     "created_at",
     "expires_at",
     "accept_token"
    ],
    "title": "InvitationCreated",
    "type": "object"
   },
   "InvitationIn": {
    "properties": {
     "email": {
      "format": "email",
      "title": "Email",
      "type": "string"
     },
     "role": {
      "enum": [
       "OWNER",
       "ADMIN",
       "ANALYST",
       "VIEWER",
       "SENDER"
      ],
      "title": "Role",
      "type": "string"
     }
    },
    "required": [
     "email",
     "role"
    ],
    "title": "InvitationIn",
    "type": "object"
   },
   "InvitationOut": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "email": {
      "title": "Email",
      "type": "string"
     },
     "expires_at": {
      "format": "date-time",
      "title": "Expires At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "role": {
      "title": "Role",
      "type": "string"
     }
    },
    "required": [
     "id",
     "email",
     "role",
     "created_at",
     "expires_at"
    ],
    "title": "InvitationOut",
    "type": "object"
   },
...
   "MeOut": {
    "properties": {
     "can_write": {
      "title": "Can Write",
      "type": "boolean"
     },
     "csrf_token": {
      "title": "Csrf Token",
      "type": "string"
     },
     "permissions": {
      "default": [],
      "items": {
       "type": "string"
      },
      "title": "Permissions",
      "type": "array"
     },
     "role": {
      "title": "Role",
      "type": "string"
     },
     "tenant": {
      "$ref": "#/components/schemas/TenantOut"
     },
     "user": {
      "$ref": "#/components/schemas/UserOut"
     }
    },
    "required": [
     "user",
     "tenant",
     "role",
     "can_write",
     "csrf_token"
    ],
    "title": "MeOut",
    "type": "object"
   },
   "MemberOut": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "email": {
      "title": "Email",
      "type": "string"
     },
     "membership_id": {
      "title": "Membership Id",
      "type": "string"
     },
     "role": {
      "title": "Role",
      "type": "string"
     },
     "user_id": {
      "title": "User Id",
      "type": "string"
     }
    },
    "required": [
     "membership_id",
     "user_id",
     "email",
     "display_name",
     "role",
     "created_at"
    ],
    "title": "MemberOut",
    "type": "object"
   },
...
   "OrgOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "org_type": {
      "title": "Org Type",
      "type": "string"
     },
     "require_2fa": {
      "title": "Require 2Fa",
      "type": "boolean"
     },
     "retention_days": {
      "title": "Retention Days",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "name",
     "org_type",
     "require_2fa",
     "retention_days"
    ],
    "title": "OrgOut",
    "type": "object"
   },
   "OrgPatch": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "org_type": {
      "anyOf": [
       {
        "enum": [
         "capacity_provider",
         "mga",
         "tpa"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Org Type"
     },
     "require_2fa": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Require 2Fa"
     }
    },
    "title": "OrgPatch",
    "type": "object"
   },
...
   "RolePatch": {
    "properties": {
     "role": {
      "enum": [
       "OWNER",
       "ADMIN",
       "ANALYST",
       "VIEWER",
       "SENDER"
      ],
      "title": "Role",
      "type": "string"
     }
    },
    "required": [
     "role"
    ],
    "title": "RolePatch",
    "type": "object"
   },
...
   "TenantOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "org_type": {
      "default": "capacity_provider",
      "title": "Org Type",
      "type": "string"
     },
     "require_2fa": {
      "default": false,
      "title": "Require 2Fa",
      "type": "boolean"
     },
     "retention_days": {
      "title": "Retention Days",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "name",
     "retention_days"
    ],
    "title": "TenantOut",
    "type": "object"
   },
   "UserOut": {
    "properties": {
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "email": {
      "title": "Email",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     }
    },
    "required": [
     "id",
     "email",
     "display_name"
    ],
    "title": "UserOut",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "TrueBind API",
  "version": "0.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/auth/invitations/accept": {
   "post": {
    "operationId": "accept_invitation_api_v1_auth_invitations_accept_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AcceptIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Accept Invitation",
    "tags": [
     "auth"
    ]
   }
  },
...
  "/api/v1/org": {
   "get": {
    "operationId": "get_org_api_v1_org_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrgOut"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Org",
    "tags": [
     "organisation"
    ]
   },
   "patch": {
    "operationId": "patch_org_api_v1_org_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OrgPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrgOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Org",
    "tags": [
     "organisation"
    ]
   }
  },
  "/api/v1/org/invitations": {
   "get": {
    "operationId": "list_invitations_api_v1_org_invitations_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/InvitationOut"
         },
         "title": "Response List Invitations Api V1 Org Invitations Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Invitations",
    "tags": [
     "organisation"
    ]
   },
   "post": {
    "operationId": "create_invitation_api_v1_org_invitations_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationCreated"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Invitation",
    "tags": [
     "organisation"
    ]
   }
  },
  "/api/v1/org/invitations/{invitation_id}": {
   "delete": {
    "operationId": "revoke_invitation_api_v1_org_invitations__invitation_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "invitation_id",
      "required": true,
      "schema": {
       "title": "Invitation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Revoke Invitation",
    "tags": [
     "organisation"
    ]
   }
  },
...
  "/api/v1/org/members/{membership_id}": {
   "delete": {
    "operationId": "remove_member_api_v1_org_members__membership_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "membership_id",
      "required": true,
      "schema": {
       "title": "Membership Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Remove Member",
    "tags": [
     "organisation"
    ]
   },
   "patch": {
    "operationId": "change_role_api_v1_org_members__membership_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "membership_id",
      "required": true,
      "schema": {
       "title": "Membership Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RolePatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MemberOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Change Role",
    "tags": [
     "organisation"
    ]
   }
  },
...
 }
}